Coordinate and Dimension Refactoring: The pull request introduces a comprehensive refactoring of how coordinates and dimensions are handled, moving towards consistent use of qreal for precise measurements and QPoint for integer-based screen positions.
Improved Button Rendering: Button painting logic has been refined, particularly for rounded corners, and repaint triggers have been expanded to ensure visual consistency across various button states. GTK button icon scaling and antialiasing behavior have also been adjusted.
Efficient Menu Button Management: The mechanism for updating menu button groups has been made more efficient by updating existing components rather than deleting and recreating them.
Enhanced Drag and Move Operations: Drag and move functionalities now consistently utilize integer QPoint coordinates, especially for X11 interactions, to improve precision and reliability.

+++ src/Button.h
@@ -44,7 +44,7 @@ public:
     Q_PROPERTY(int animationDuration READ animationDuration WRITE setAnimationDuration NOTIFY animationDurationChanged)
     Q_PROPERTY(qreal opacity READ opacity WRITE setOpacity NOTIFY opacityChanged)
     Q_PROPERTY(qreal transitionValue READ transitionValue WRITE setTransitionValue NOTIFY transitionValueChanged)
-    // Q_PROPERTY(QMarginsF* padding READ padding NOTIFY paddingChanged)
+    //Q_PROPERTY(QMarginsF* padding READ padding NOTIFY paddingChanged)
 
     // Passed to DecorationButtonGroup in Decoration
     static KDecoration3::DecorationButton *create(KDecoration3::DecorationButtonType type, KDecoration3::Decoration *decoration, QObject *parent = nullptr);
@@ -58,8 +58,10 @@ public:
     void paint(QPainter *painter, const QRectF &repaintRegion) override;
     virtual void paintIcon(QPainter *painter, const QRectF &iconRect, const qreal);
 
+
     virtual void updateSize(qreal contentWidth, qreal contentHeight);
-    virtual void setHeight(int buttonHeight);
+    virtual void setHeight(qreal buttonHeight);
+
 
     void forceUnpress();
